Move SSA related code to CodeGen_SSA
